Moved CandidatesScreen into separate file
